Revert commit 1fde398d1 and follow-on patches.
This reverts 1fde398d1 (Allow planner to use Merge Append to
efficiently implement UNION) as well as the follow-on commits
e13613fb8, bf66b4e6f, 2d990f808.  In addition to those, a40a1cb40
had to be removed then re-applied in a different place, because
1fde398d1 moved the relevant code.

The reason for this last-minute thrashing is that depesz found a
case in which the patched code creates a completely wrong plan
that silently gives incorrect query results.  It's unclear what
the cause is or how many cases are affected, but with beta1 wrap
staring us in the face, there's no time for closer investigation.
After we figure that out, we can decide whether to un-revert this
for beta2 or hold it for v18.

Discussion: https://postgr.es/m/Zktzf926vslR35Fv@depesz.com
(also some private discussion among pgsql-release)